+++ src/components/CustomComponents_FR/index.js
@@ -1,2 +1,3 @@
 import * as AllCustomComponentsFR from './CustomComponents_FR';
 export { AllCustomComponentsFR };
+export { Grammar1Body, Grammar2Body } from './CustomComponents_FR';